refactor: deduplicate Vector3 and TaskMessage into nextgsim-common
- Move Vector3 (f64) from nextgsim-isac and nextgsim-nwdaf to
  nextgsim-common::types with distance_to() and magnitude() methods
- Move TaskMessage<T> from nextgsim-gnb and nextgsim-ue to
  nextgsim-common::types with message(), shutdown(), is_shutdown()
- Re-export both types from original crates for API compatibility
- RLS Vector3 (i32) left unchanged — semantically different type
